move geog_point tests
