+++ python/sedonadb/tests/functions/test_functions.py
@@ -2404,31 +2404,6 @@ def test_st_reverse(eng, geom, expected):
         (10, -1.5, "POINT (10 -1.5)"),
     ],
 )
-def test_st_geogpoint(eng, x, y, expected):
-    eng = eng.create_or_skip()
-    if eng == SedonaDB:
-        eng.assert_query_result(
-            f"SELECT ST_GeogPoint({val_or_null(x)}, {val_or_null(y)})", expected
-        )
-    else:
-        eng.assert_query_result(
-            f"SELECT ST_Point({val_or_null(x)}, {val_or_null(y)}) as geography",
-            expected,
-        )
-
-
-@pytest.mark.parametrize("eng", [SedonaDB, PostGIS])
-@pytest.mark.parametrize(
-    ("x", "y", "expected"),
-    [
-        (None, None, None),
-        (1, None, None),
-        (None, 1, None),
-        (1, 1, "POINT (1 1)"),
-        (1.0, 1.0, "POINT (1 1)"),
-        (10, -1.5, "POINT (10 -1.5)"),
-    ],
-)
 def test_st_point(eng, x, y, expected):
     eng = eng.create_or_skip()
     eng.assert_query_result(
