+++ python/sedonadb/tests/geography/bigquery_cache.yml
@@ -31,3 +31,9 @@ results:
     SELECT ST_Area(ST_GeogFromText('POINT EMPTY')): /////3gAAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAUAAAAEAAUAAgABgAHAAwAAAAQABAAAAAAAAEDEAAAABwAAAAEAAAAAAAAAAMAAABmMF8AAAAGAAgABgAGAAAAAAACAAAAAAD/////iAAAABQAAAAAAAAADAAWAAYABQAIAAwADAAAAAADBAAYAAAACAAAAAAAAAAAAAoAGAAMAAQACAAKAAAAPAAAABAAAAABAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAQAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAP////8AAAAA
     SELECT ST_Area(ST_GeogFromText('POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))')): /////3gAAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAUAAAAEAAUAAgABgAHAAwAAAAQABAAAAAAAAEDEAAAABwAAAAEAAAAAAAAAAMAAABmMF8AAAAGAAgABgAGAAAAAAACAAAAAAD/////iAAAABQAAAAAAAAADAAWAAYABQAIAAwADAAAAAADBAAYAAAACAAAAAAAAAAAAAoAGAAMAAQACAAKAAAAPAAAABAAAAABAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAQAAAAEAAAAAAAAAAAAAAAAAAACBnLjOoQcHQv////8AAAAA
     SELECT ST_Area(ST_GeogFromText('POLYGON EMPTY')): /////3gAAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAUAAAAEAAUAAgABgAHAAwAAAAQABAAAAAAAAEDEAAAABwAAAAEAAAAAAAAAAMAAABmMF8AAAAGAAgABgAGAAAAAAACAAAAAAD/////iAAAABQAAAAAAAAADAAWAAYABQAIAAwADAAAAAADBAAYAAAACAAAAAAAAAAAAAoAGAAMAAQACAAKAAAAPAAAABAAAAABAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAQAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAP////8AAAAA
+    SELECT ST_GeogPoint(1, 1): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABgAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAAAAAACAAAAAAAAAAKAAAAAAAAAAAAAAABAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAKAAAAUE9JTlQoMSAxKQAAAAAAAP////8AAAAA
+    SELECT ST_GeogPoint(1, NULL): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABAAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAQAAAAAAAAAIAAAAAAAAAAgAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAQAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AAAAAA==
+    SELECT ST_GeogPoint(1.0, 1.0): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABgAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAAAAAACAAAAAAAAAAKAAAAAAAAAAAAAAABAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAKAAAAUE9JTlQoMSAxKQAAAAAAAP////8AAAAA
+    SELECT ST_GeogPoint(10, -1.5): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABgAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAAAAAACAAAAAAAAAAOAAAAAAAAAAAAAAABAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAOAAAAUE9JTlQoMTAgLTEuNSkAAP////8AAAAA
+    SELECT ST_GeogPoint(NULL, 1): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABAAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAQAAAAAAAAAIAAAAAAAAAAgAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAQAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AAAAAA==
+    SELECT ST_GeogPoint(NULL, NULL): /////xgBAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAEAAAAYAAAAAAASABgACAAGAAcADAAAABAAFAASAAAAAAABBRQAAAC8AAAACAAAABAAAAAAAAAAAwAAAGYwXwACAAAAVAAAAAQAAAC8////IAAAAAQAAAATAAAAeyJlbmNvZGluZyI6ICJXS1QifQAYAAAAQVJST1c6ZXh0ZW5zaW9uOm1ldGFkYXRhAAAAAAgADAAEAAgACAAAACgAAAAEAAAAGAAAAGdvb2dsZTpzcWxUeXBlOmdlb2dyYXBoeQAAAAAUAAAAQVJST1c6ZXh0ZW5zaW9uOm5hbWUAAAAABAAEAAQAAAAAAAAA/////5gAAAAUAAAAAAAAAAwAFgAGAAUACAAMAAwAAAAAAwQAGAAAABAAAAAAAAAAAAAKABgADAAEAAgACgAAAEwAAAAQAAAAAQAAAAAAAAAAAAAAAwAAAAAAAAAAAAAAAQAAAAAAAAAIAAAAAAAAAAgAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAQAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AAAAAA==
